+++ lib/src/controls/controls.dart
@@ -1,3 +1,4 @@
 export 'flutter_deck_controls.dart';
+export 'flutter_deck_controls_configuration.dart';
 export 'flutter_deck_controls_listener.dart';
 export 'flutter_deck_controls_notifier.dart';