+++ lib/src/widgets/internal/internal.dart
@@ -1,3 +1,2 @@
-export 'controls/controls.dart';
 export 'drawer/drawer.dart';
 export 'marker/marker.dart';